pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cascade"
version = "0.1.0"
description = "HEP Workflow Orchestrator"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "zhipuai>=2.1.5.20230904",
    "python-dotenv>=1.0.1",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.43",
    "click>=8.1.7"
]

[project.scripts]
cascade = "cascade.cli.main:main"
